refactor: reorganize built files to follow a scoped packages structure
